fix: solve inc/dec redundancy problem
팔로우, 팔로우 취소 시에 클라이언트 사이드에서 상태가 변하지 않아 중복된 요청에 대하여 실제 팔로우/팔로우 취소 여부와 관계없이 반복적으로 팔로워,팔로잉 수가 증감되는 문제를 팔로우 로직 이전에 존재하는지/존재하지 않는지 확인하는 것으로 해결하였습니다.

+++ app/src/main/java/com/delivalue/tidings/domain/follow/service/FollowService.java
@@ -26,14 +26,17 @@ public class FollowService {
     @Transactional
     public void addFollowUser(String followerId, String followingUserPublicId) {
         Member followingUser = this.memberRepository.findByPublicId(followingUserPublicId);
-        if(followingUser == null) throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        if(followingUser == null || followerId.equals(followingUser.getId())) throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
 
         FollowId id = new FollowId(followingUser.getId(), followerId);
         Follow followEntity = Follow.builder().id(id).createdAt(LocalDateTime.now(ZoneId.of("Asia/Seoul"))).build();
-        this.followRepository.save(followEntity);
+        boolean exists = this.followRepository.existsById(id);
+        if(!exists) {
+            this.followRepository.save(followEntity);
 
-        this.memberRepository.increaseFollowerCount(followingUser.getId());
-        this.memberRepository.increaseFollowingCount(followerId);
+            this.memberRepository.increaseFollowerCount(followingUser.getId());
+            this.memberRepository.increaseFollowingCount(followerId);
+        }
     }
 
     @Transactional
@@ -42,10 +45,13 @@ public class FollowService {
         if(followingUser == null) throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
 
         FollowId id = new FollowId(followingUser.getId(), followerId);
-        this.followRepository.deleteById(id);
+        boolean exists = this.followRepository.existsById(id);
+        if(exists) {
+            this.followRepository.deleteById(id);
 
-        this.memberRepository.decreaseFollowerCount(followingUser.getId());
-        this.memberRepository.decreaseFollowingCount(followerId);
+            this.memberRepository.decreaseFollowerCount(followingUser.getId());
+            this.memberRepository.decreaseFollowingCount(followerId);
+        }
     }
 
     public List<ProfileResponse> getFollowingList(String publicId) {
